type erasure helps make a universal parser
