+++ include/mechanism_configuration/parser_base.hpp
@@ -3,6 +3,7 @@
 #include <yaml-cpp/yaml.h>
 
 #include <mechanism_configuration/mechanism.hpp>
+#include <mechanism_configuration/load_node.hpp>
 
 #include <filesystem>
 #include <iostream>
@@ -13,7 +14,7 @@ namespace mechanism_configuration
 {
   using GlobalMechanism = ::mechanism_configuration::Mechanism;
 
-  template<typename MechanismType>
+  template<typename MechanismType = GlobalMechanism>
   class ParserBase
   {
    public:
@@ -42,38 +43,5 @@ namespace mechanism_configuration
     /// @param source A YAML node, file path, or string representing a file path
     /// @return An optional containing the parsed mechanism wrapped in a unique pointer
     virtual std::optional<std::unique_ptr<GlobalMechanism>> TryParse(const YAML::Node& node) = 0;
-
-   protected:
-    /// @brief Helper to load YAML node from a file path, string, or YAML::Node
-    /// @param source A file path, string, or YAML::Node
-    /// @return A YAML node
-    template<typename T>
-    YAML::Node LoadNode(const T& source)
-    {
-      if constexpr (std::is_same_v<std::decay_t<T>, YAML::Node>)
-      {
-        return source;
-      }
-      else if constexpr (std::is_same_v<std::decay_t<T>, std::string> || std::is_same_v<std::decay_t<T>, std::filesystem::path>)
-      {
-        // check if the file exists
-        if (std::filesystem::exists(source))
-        {
-          return YAML::LoadFile(source);
-        }
-        return YAML::Node();
-      }
-      else
-      {
-        static_assert(always_false<T>::value, "Unsupported type for LoadNode");
-      }
-    }
-
-   private:
-    // Helper for static_assert to provide meaningful error messages
-    template<typename>
-    struct always_false : std::false_type
-    {
-    };
   };
 }  // namespace mechanism_configuration
